requirements.txt
fastapi==0.103.0
uvicorn==0.23.2
httpx==0.24.1
qdrant-client==1.4.0
openai==0.27.9
langchain==0.0.274
python-dotenv==1.0.0
pypdf==3.15.4
unstructured==0.10.6
pdf2image==1.16.3
tiktoken==0.4.0